{routes-spotify-merged} working on making this a MVC-accurate structure.
